IN-41 refactor: add some fixes
