+++ src/features/auth/model/types/auth.ts
@@ -10,11 +10,33 @@ export type RegistrationEmailResendingInput = {
   email: string
 }
 
-export type recoveryCodeCheckRequest = {
+export type RecoveryCodeCheckRequest = {
   recoveryCode: string
 }
 
 export type NewPasswordRequest = {
   newPassword: string
   recoveryCode: string
+}
+
+export type LoginRequest = {
+  email: string
+  password: string
+}
+
+export type PasswordRecoveryRequest = {
+  // baseUrl: string
+  email: string
+  recaptcha: null | string
+}
+
+export type LoginResponse = {
+  accessToken: string
+}
+
+export type MeResponse = {
+  email: string
+  isBlocked: boolean
+  userId: number
+  userName: string
 }
