+++ src/features/profile/api/profileApi.ts
@@ -23,13 +23,13 @@ export const profileApi = baseApi.injectEndpoints({
       },
       query: () => ({
         method: 'DELETE',
-        url: apiEndpoints.public.user.deleteAvatar,
+        url: apiEndpoints.profile.avatar,
       }),
     }),
     getProfile: builder.query<GetProfileResponse, void>({
       query: () => ({
         method: 'GET',
-        url: apiEndpoints.public.user.profile,
+        url: apiEndpoints.profile.profile,
       }),
     }),
     uploadAvatar: builder.mutation<AddAvatarResponse, FormData>({
@@ -56,7 +56,7 @@ export const profileApi = baseApi.injectEndpoints({
       query: FormData => ({
         body: FormData,
         method: 'POST',
-        url: apiEndpoints.public.user.uploadAvatar,
+        url: apiEndpoints.profile.avatar,
       }),
     }),
   }),
